Remove upper-binding for SQLAlchemy (#24819)
There was a problem with custom classes for SQLAlchemy that
prevented it to work on MySQL. This PR removes the SQLAlchemy
upper binding.

This has been added as an issue in:

https://github.com/sqlalchemy/sqlalchemy/issues/7660

But apparently it's been fixed in one of the more recent SQLAlchemy
releases.

(cherry picked from commit 0853d60f16b45472ebac83bcb4fb6de112c90d6b)